requirements.txt
Flask
pandas
pillow
gunicorn
openpyxl